TST: Suppress Series constructor warning (#48804)

+++ pandas/tests/groupby/test_size.py
@@ -33,7 +33,7 @@ def test_size_axis_1(df, axis_1, by, sort, dropna):
     counts = {key: sum(value == key for value in by) for key in dict.fromkeys(by)}
     if dropna:
         counts = {key: value for key, value in counts.items() if key is not None}
-    expected = Series(counts)
+    expected = Series(counts, dtype="int64")
     if sort:
         expected = expected.sort_index()
     grouped = df.groupby(by=by, axis=axis_1, sort=sort, dropna=dropna)
